+++ src/GRA.Data/Model/Program.cs
@@ -13,8 +13,15 @@ namespace GRA.Data.Model
         public string Name { get; set; }
         [Required]
         public int AchieverPointAmount { get; set; }
+
         public int? AchieverBadgeId { get; set; }
+        [MaxLength(255)]
+        public string AchieverBadgeName { get; set; }
+
         public int? JoinBadgeId { get; set; }
+        [MaxLength(255)]
+        public string JoinBadgeName { get; set; }
+
         [Required]
         public bool AskAge { get; set; }
         [Required]